[refactor/#47] RegisterEnterDate, RegisterEnterTime -> RegisterEnterDateTime
